+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **[FIX]** Fix `MediaSizeProvider` first render state
+- **[NEW]** New serverSideMediaSize attribute for `MediaSizeProvider`
 [...]
 
 # v45.1.0 (20/01/2021)